Move template literal logic into its extractor.
No need to separate logic here. It makes the code less readable.
